requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.0.0
numpy>=1.21.0
pyyaml>=5.1
matplotlib>=3.3.0
scikit-learn>=0.24.0
fastapi>=0.68.0
requests>=2.28.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
SQLAlchemy>=2.0.0
google-generativeai>=0.3.0
tensorflow>=2.16.1
googletrans>=4.0.2
deep-translator>=1.11.4
itsdangerous>=2.2.0
python-dotenv>=1.0.1